Added support for sorting data
Earlier there was an attempt to add support for
data sorting, which was implemented in the way
that every column will have an associated radio
button that can be turned on to use that column
for sorting. Now there is a single drop down
where one column can be selected, and that is used
to sort the data.